Version 2.0.2

+++ Demo/PDFTouch.framework/Versions/A/Headers/YLCompositeFont.h
@@ -20,6 +20,6 @@
 
 @interface YLCompositeFont : YLFont
 
-@property (nonatomic, readonly) NSUInteger defaultWidth;
+@property (nonatomic, readonly) CGFloat defaultWidth;
 
 @end
